moved data generator functions to their own module
